restructuring the views. Changes to be committed: 	deleted:    app/views.py 	new file:   app/views/__init__.py 	new file:   app/views/auth.py 	new file:   app/views/questions.py
